Added ability to deal with symmetric TPCs.
A function has been added to check if a detector is symmetryc or not.
The fiduciality check now vetoes also the central cathode region.
Delayed hits are shifted in Z in the opposite direction to the drift,
therefore in different directions depending on the side of the TPC.
Tests for these two functionalities in symmetric TPCs have been added.