Added an S1 filter at Montecarlo Energy Level.
Added a single S1 filter based on the time of MC hits.
In practice it is only applied to Bi214 events as they are the only
ones able to have more than 1 S1s, avoiding the extra computing time
associated.
All S1s must be in a period corresponding to the recorded time of a
single event, and the energy of the associated hits must be higher
than a threshold.
Tests for this new functionality have been added.